requirements.txt
pydantic==1.9.0
pymongo[srv]==4.0.1
python-telegram-bot==13.9
pytz==2021.3